refactor(setupmockserver): refactoring mock server setup file
