+++ src/pages/bag/Bag.spec.tsx
@@ -1,13 +1,10 @@
 import React from 'react'
-import { setupServer } from 'msw/node'
 import { waitFor } from '@testing-library/react'
-import { rest } from 'msw'
 import * as router from 'react-router'
 import { render, fireEvent } from '../../test-utils'
 import { Bag } from './Bag'
-import { bagResponseMock } from '../../../mock/bagResponse'
-import { belezaNaWebApi } from '../../api'
 import { BagProvider } from '../../contexts/BagContext'
+import { setupMockServer } from '../../../mock/msw/setupMockServer'
 
 const navigate = jest.fn()
 
@@ -15,16 +12,7 @@ beforeEach(() => {
   jest.spyOn(router, 'useNavigate').mockImplementation(() => navigate)
 })
 
-const baseURL = belezaNaWebApi.getUri()
-
-const server = setupServer(
-  rest.get(`${baseURL}5b15c4923100004a006f3c07`, (req, res, ctx) =>
-    res(ctx.status(200), ctx.json(bagResponseMock))
-  )
-)
-
-beforeAll(() => server.listen())
-afterAll(() => server.close())
+setupMockServer()
 
 const renderComponent = () =>
   render(
